requirements.txt
# Framework
Django==1.7

# Libs
psycopg2==2.5.4

# Django-specific libraries
django-braces==1.4.0
django-crispy-forms==1.4.0
django-extensions==1.4.0